service_manager/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
pydantic
PyJWT
passlib[bcrypt]
pydantic[email]
requests
firebase-admin>=6.5.0
python-dotenv
pandas
openpyxl  # Needed for reading Excel files

common_utils @ git+https://github.com/YouthStack-Dev/MLT_backend_fastapi.git@feat/jwt-enhancements#subdirectory=common_utils_package
# pip install git+https://github.com/YouthStack-Dev/MLT_backend_fastapi.git#subdirectory=common_utils

